Move Geoserver names classes to geoserver package
